tests: remove duplicated unit tests
'test_provider.py' and 'test_providers.py' had duplicated tests.
This was an artifact of commit c7a49bac29, which combined the same
tests from 'test_lxd.py' and 'test_multipass.py'.
